+++ client/internal/pkg/sshcert/sshcert.go
@@ -1,7 +1,8 @@
+// The sshcert package provides a convenient function for parsing a
+// SSH certificate.
 package sshcert
 
 import (
-	"crypto/ecdsa"
 	"fmt"
 
 	"golang.org/x/crypto/ssh"
@@ -26,18 +27,4 @@ func ParseCert(certBytes []byte) (*ssh.Certificate, error) {
 	}
 
 	return cert, nil
-}
-
-func ParseKey(pemBytes []byte) (*ecdsa.PrivateKey, error) {
-	privateKey, err := ssh.ParseRawPrivateKey(pemBytes)
-	if err != nil {
-		return nil, fmt.Errorf("could not parse private key file: %w", err)
-	}
-
-	ecdsaKey, ok := privateKey.(*ecdsa.PrivateKey)
-	if !ok {
-		return nil, fmt.Errorf("private key is not an ECDSA key; its type is %T", privateKey)
-	}
-
-	return ecdsaKey, nil
 }
