+++ client/internal/pkg/sshkey/sshkey.go
@@ -1,3 +1,5 @@
+// The sshkey provides a simple way to generate a 256-bit ECDSA private key
+// for SSH.
 package sshkey
 
 import (
@@ -10,6 +12,11 @@ import (
 	"golang.org/x/crypto/ssh"
 )
 
+// GenerateKey will generate an OpenSSH ECDSA private key using
+// the P-256 elliptic curve.
+//
+// The resulting key is returned as a byte slice in OpenSSH PEM
+// format.
 func GenerateKey(comment string) ([]byte, error) {
 	// generate ECDSA key
 	key, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
@@ -29,4 +36,25 @@ func GenerateKey(comment string) ([]byte, error) {
 	}
 
 	return pemBytes, nil
+}
+
+// ParseKey will parse the provided byte slice (in OpenSSH ECDSA Private Key format)
+// and return an [*ecdsa.PrivateKey].
+//
+// Any parsing errors will result in a nil [*ecdsa.PrivateKey] returned along
+// with the error.
+//
+// Only ECDSA format private keys are supported by this function.
+func ParseKey(pemBytes []byte) (*ecdsa.PrivateKey, error) {
+	privateKey, err := ssh.ParseRawPrivateKey(pemBytes)
+	if err != nil {
+		return nil, fmt.Errorf("could not parse private key file: %w", err)
+	}
+
+	ecdsaKey, ok := privateKey.(*ecdsa.PrivateKey)
+	if !ok {
+		return nil, fmt.Errorf("private key is not an ECDSA key; its type is %T", privateKey)
+	}
+
+	return ecdsaKey, nil
 }
